refactor: move updateInteractionNameWithUriVariablePattern to core
